refactor(deposit): update deposit entity (#53)
