--- src/main/java/com/remittance/deposit/model/DepositStatus.java
@@ -1,7 +1,0 @@
-package com.remittance.deposit.model;
-
-public enum DepositStatus {
-    PENDING,
-    CONFIRMED,
-    FAILED,
-}
